Ehcache base class
