+++ core/src/main/java/org/ehcache/core/Ehcache.java
@@ -16,35 +16,20 @@
 
 package org.ehcache.core;
 
-import java.time.Duration;
-import java.util.AbstractMap;
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Collections;
-import java.util.EnumMap;
 import java.util.HashMap;
-import java.util.Iterator;
 import java.util.LinkedHashMap;
-import java.util.LinkedHashSet;
 import java.util.Map;
-import java.util.NoSuchElementException;
 import java.util.Set;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
-import java.util.concurrent.atomic.LongAdder;
 import java.util.function.BiFunction;
 import java.util.function.Function;
 import java.util.function.Supplier;
 
 import org.ehcache.Cache;
-import org.ehcache.Status;
 import org.ehcache.config.CacheConfiguration;
-import org.ehcache.config.CacheRuntimeConfiguration;
 import org.ehcache.core.events.CacheEventDispatcher;
-import org.ehcache.core.internal.resilience.LoggingRobustResilienceStrategy;
-import org.ehcache.core.internal.resilience.RecoveryCache;
-import org.ehcache.core.internal.resilience.ResilienceStrategy;
-import org.ehcache.core.spi.LifeCycled;
 import org.ehcache.core.spi.store.Store;
 import org.ehcache.core.spi.store.Store.PutStatus;
 import org.ehcache.core.spi.store.Store.RemoveStatus;
@@ -52,7 +37,6 @@ import org.ehcache.core.spi.store.Store.ReplaceStatus;
 import org.ehcache.core.spi.store.Store.ValueHolder;
 import org.ehcache.core.spi.store.StoreAccessException;
 import org.ehcache.core.statistics.BulkOps;
-import org.ehcache.core.statistics.CacheOperationOutcomes.ClearOutcome;
 import org.ehcache.core.statistics.CacheOperationOutcomes.ConditionalRemoveOutcome;
 import org.ehcache.core.statistics.CacheOperationOutcomes.GetAllOutcome;
 import org.ehcache.core.statistics.CacheOperationOutcomes.GetOutcome;
@@ -67,12 +51,6 @@ import org.ehcache.spi.loaderwriter.BulkCacheLoadingException;
 import org.ehcache.spi.loaderwriter.BulkCacheWritingException;
 import org.ehcache.spi.loaderwriter.CacheLoaderWriter;
 import org.slf4j.Logger;
-import org.terracotta.statistics.StatisticsManager;
-import org.terracotta.statistics.observer.OperationObserver;
-
-import static org.ehcache.core.exceptions.ExceptionFactory.newCacheLoadingException;
-import static org.ehcache.core.internal.util.ValueSuppliers.supplierOf;
-import static org.terracotta.statistics.StatisticBuilder.operation;
 
 /**
  * Implementation of the {@link Cache} interface when no {@link CacheLoaderWriter} is involved.
@@ -82,27 +60,7 @@ import static org.terracotta.statistics.StatisticBuilder.operation;
  *
  * @see EhcacheWithLoaderWriter
  */
-public class Ehcache<K, V> implements InternalCache<K, V> {
-
-  private final StatusTransitioner statusTransitioner;
-
-  private final Store<K, V> store;
-  private final ResilienceStrategy<K, V> resilienceStrategy;
-  private final EhcacheRuntimeConfiguration<K, V> runtimeConfiguration;
-  private final Jsr107CacheImpl jsr107Cache;
-  protected final Logger logger;
-
-  private final OperationObserver<GetOutcome> getObserver = operation(GetOutcome.class).named("get").of(this).tag("cache").build();
-  private final OperationObserver<GetAllOutcome> getAllObserver = operation(GetAllOutcome.class).named("getAll").of(this).tag("cache").build();
-  private final OperationObserver<PutOutcome> putObserver = operation(PutOutcome.class).named("put").of(this).tag("cache").build();
-  private final OperationObserver<PutAllOutcome> putAllObserver = operation(PutAllOutcome.class).named("putAll").of(this).tag("cache").build();
-  private final OperationObserver<RemoveOutcome> removeObserver = operation(RemoveOutcome.class).named("remove").of(this).tag("cache").build();
-  private final OperationObserver<RemoveAllOutcome> removeAllObserver = operation(RemoveAllOutcome.class).named("removeAll").of(this).tag("cache").build();
-  private final OperationObserver<ConditionalRemoveOutcome> conditionalRemoveObserver = operation(ConditionalRemoveOutcome.class).named("conditionalRemove").of(this).tag("cache").build();
-  private final OperationObserver<PutIfAbsentOutcome> putIfAbsentObserver = operation(PutIfAbsentOutcome.class).named("putIfAbsent").of(this).tag("cache").build();
-  private final OperationObserver<ReplaceOutcome> replaceObserver = operation(ReplaceOutcome.class).named("replace").of(this).tag("cache").build();
-  private final Map<BulkOps, LongAdder> bulkMethodEntries = new EnumMap<>(BulkOps.class);
-  private final OperationObserver<ClearOutcome> clearObserver = operation(ClearOutcome.class).named("clear").of(this).tag("cache").build();
+public class Ehcache<K, V> extends EhcacheBase<K, V> {
 
   /**
    * Creates a new {@code Ehcache} based on the provided parameters.
@@ -117,43 +75,8 @@ public class Ehcache<K, V> implements InternalCache<K, V> {
   }
 
   Ehcache(EhcacheRuntimeConfiguration<K, V> runtimeConfiguration, Store<K, V> store,
-            CacheEventDispatcher<K, V> eventDispatcher, Logger logger, StatusTransitioner statusTransitioner) {
-    this.store = store;
-    runtimeConfiguration.addCacheConfigurationListener(store.getConfigurationChangeListeners());
-    StatisticsManager.associate(store).withParent(this);
-
-    if (store instanceof RecoveryCache) {
-      this.resilienceStrategy = new LoggingRobustResilienceStrategy<>(castToRecoveryCache(store));
-    } else {
-      this.resilienceStrategy = new LoggingRobustResilienceStrategy<>(recoveryCache(store));
-    }
-
-    this.runtimeConfiguration = runtimeConfiguration;
-    runtimeConfiguration.addCacheConfigurationListener(eventDispatcher.getConfigurationChangeListeners());
-    this.jsr107Cache = new Jsr107CacheImpl();
-
-    this.logger=logger;
-    this.statusTransitioner = statusTransitioner;
-    for (BulkOps bulkOp : BulkOps.values()) {
-      bulkMethodEntries.put(bulkOp, new LongAdder());
-    }
-  }
-
-  /**
-   * {@inheritDoc}
-   */
-  @Override
-  public Map<BulkOps, LongAdder> getBulkMethodEntries() {
-    return bulkMethodEntries;
-  }
-
-  @SuppressWarnings("unchecked")
-  private RecoveryCache<K> castToRecoveryCache(Store<K, V> store) {
-    return (RecoveryCache<K>) store;
-  }
-
-  private V getNoLoader(K key) {
-    return get(key);
+          CacheEventDispatcher<K, V> eventDispatcher, Logger logger, StatusTransitioner statusTransitioner) {
+    super(runtimeConfiguration, store, eventDispatcher, logger, statusTransitioner);
   }
 
   /**
@@ -161,28 +84,7 @@ public class Ehcache<K, V> implements InternalCache<K, V> {
    */
   @Override
   public V get(final K key) {
-    getObserver.begin();
-    statusTransitioner.checkAvailable();
-    checkNonNull(key);
-
-    try {
-      final Store.ValueHolder<V> valueHolder = store.get(key);
-
-      // Check for expiry first
-      if (valueHolder == null) {
-        getObserver.end(GetOutcome.MISS);
-        return null;
-      } else {
-        getObserver.end(GetOutcome.HIT);
-        return valueHolder.value();
-      }
-    } catch (StoreAccessException e) {
-      try {
-        return resilienceStrategy.getFailure(key, e);
-      } finally {
-        getObserver.end(GetOutcome.FAILURE);
-      }
-    }
+    return getNoLoader(key);
   }
 
   /**
@@ -215,34 +117,7 @@ public class Ehcache<K, V> implements InternalCache<K, V> {
     }
   }
 
-  private boolean newValueAlreadyExpired(K key, V oldValue, V newValue) {
-    return newValueAlreadyExpired(logger, runtimeConfiguration.getExpiryPolicy(), key, oldValue, newValue);
-  }
-
-  /**
-   * {@inheritDoc}
-   */
-  @Override
-  public boolean containsKey(final K key) {
-    statusTransitioner.checkAvailable();
-    checkNonNull(key);
-    try {
-      return store.containsKey(key);
-    } catch (StoreAccessException e) {
-      return resilienceStrategy.containsKeyFailure(key, e);
-    }
-  }
-
-  /**
-   * {@inheritDoc}
-   */
-  @Override
-  public void remove(K key) {
-    removeInternal(key); // ignore return value;
-  }
-
-
-  private boolean removeInternal(final K key) {
+  protected boolean removeInternal(final K key) {
     removeObserver.begin();
     statusTransitioner.checkAvailable();
     checkNonNull(key);
@@ -266,40 +141,7 @@ public class Ehcache<K, V> implements InternalCache<K, V> {
     return removed;
   }
 
-  /**
-   * {@inheritDoc}
-   */
-  @Override
-  public void clear() {
-    this.clearObserver.begin();
-    statusTransitioner.checkAvailable();
-    try {
-      store.clear();
-        this.clearObserver.end(ClearOutcome.SUCCESS);
-    } catch (StoreAccessException e) {
-      this.clearObserver.end(ClearOutcome.FAILURE);
-      resilienceStrategy.clearFailure(e);
-    }
-  }
-
-  /**
-   * {@inheritDoc}
-   */
-  @Override
-  public Iterator<Entry<K, V>> iterator() {
-    statusTransitioner.checkAvailable();
-    return new CacheEntryIterator(false);
-  }
-
-  /**
-   * {@inheritDoc}
-   */
-  @Override
-  public Map<K, V> getAll(Set<? extends K> keys) throws BulkCacheLoadingException {
-    return getAllInternal(keys, true);
-  }
-
-  private Map<K, V> getAllInternal(Set<? extends K> keys, boolean includeNulls) throws BulkCacheLoadingException {
+  protected Map<K, V> getAllInternal(Set<? extends K> keys, boolean includeNulls) throws BulkCacheLoadingException {
     getAllObserver.begin();
     statusTransitioner.checkAvailable();
     checkNonNullContent(keys);
@@ -335,14 +177,6 @@ public class Ehcache<K, V> implements InternalCache<K, V> {
         getAllObserver.end(GetAllOutcome.FAILURE);
       }
     }
-  }
-
-  LinkedHashSet<Map.Entry<? extends K, ? extends V>> nullValuesForKeys(final Iterable<? extends K> keys) {
-    final LinkedHashSet<Map.Entry<? extends K, ? extends V>> entries = new LinkedHashSet<>();
-    for (K key : keys) {
-      entries.add(new AbstractMap.SimpleEntry<>(key, null));
-    }
-    return entries;
   }
 
   /**
@@ -426,10 +260,9 @@ public class Ehcache<K, V> implements InternalCache<K, V> {
     statusTransitioner.checkAvailable();
     checkNonNull(key, value);
 
-    boolean absent = false;
     try {
       ValueHolder<V> inCache = store.putIfAbsent(key, value);
-      absent = (inCache == null);
+      boolean absent = (inCache == null);
       if (absent) {
         putIfAbsentObserver.end(PutIfAbsentOutcome.PUT);
         return null;
@@ -439,7 +272,7 @@ public class Ehcache<K, V> implements InternalCache<K, V> {
       }
     } catch (StoreAccessException e) {
       try {
-        return resilienceStrategy.putIfAbsentFailure(key, value, null, e, absent);
+        return resilienceStrategy.putIfAbsentFailure(key, value, null, e, false); // FIXME: We can't know if it's absent or not
       } finally {
         putIfAbsentObserver.end(PutIfAbsentOutcome.FAILURE);
       }
@@ -454,11 +287,10 @@ public class Ehcache<K, V> implements InternalCache<K, V> {
     conditionalRemoveObserver.begin();
     statusTransitioner.checkAvailable();
     checkNonNull(key, value);
-    RemoveStatus status = null;
     boolean removed = false;
 
     try {
-      status = store.remove(key, value);
+      RemoveStatus status = store.remove(key, value);
       switch (status) {
       case REMOVED:
         removed = true;
@@ -475,7 +307,7 @@ public class Ehcache<K, V> implements InternalCache<K, V> {
       }
     } catch (StoreAccessException e) {
       try {
-        return resilienceStrategy.removeFailure(key, value, e, removed);
+        return resilienceStrategy.removeFailure(key, value, e, false); // FIXME: We can't know if it's removed or not
       } finally {
         conditionalRemoveObserver.end(ConditionalRemoveOutcome.FAILURE);
       }
@@ -518,11 +350,10 @@ public class Ehcache<K, V> implements InternalCache<K, V> {
     statusTransitioner.checkAvailable();
     checkNonNull(key, oldValue, newValue);
 
-    ReplaceStatus status = null;
     boolean success = false;
 
     try {
-      status = store.replace(key, oldValue, newValue);
+      ReplaceStatus status = store.replace(key, oldValue, newValue);
       switch (status) {
       case HIT:
         success = true;
@@ -541,86 +372,16 @@ public class Ehcache<K, V> implements InternalCache<K, V> {
       return success;
     } catch (StoreAccessException e) {
       try {
-        return resilienceStrategy.replaceFailure(key, oldValue, newValue, e, success);
+        return resilienceStrategy.replaceFailure(key, oldValue, newValue, e, false); // FIXME: We can't know if there was a match
       } finally {
         replaceObserver.end(ReplaceOutcome.FAILURE);
       }
     }
   }
 
-  /**
-   * {@inheritDoc}
-   */
   @Override
-  public CacheRuntimeConfiguration<K, V> getRuntimeConfiguration() {
-    return runtimeConfiguration;
-  }
-
-  /**
-   * {@inheritDoc}
-   */
-  @Override
-  public void init() {
-    statusTransitioner.init().succeeded();
-  }
-
-  /**
-   * {@inheritDoc}
-   */
-  @Override
-  public void close() {
-    statusTransitioner.close().succeeded();
-  }
-
-  /**
-   * {@inheritDoc}
-   */
-  @Override
-  public Status getStatus() {
-    return statusTransitioner.currentStatus();
-  }
-
-  /**
-   * {@inheritDoc}
-   */
-  @Override
-  public void addHook(LifeCycled hook) {
-    statusTransitioner.addHook(hook);
-  }
-
-  void removeHook(LifeCycled hook) {
-    statusTransitioner.removeHook(hook);
-  }
-
-  private static void checkNonNull(Object thing) {
-    if(thing == null) {
-      throw new NullPointerException();
-    }
-  }
-
-  private static void checkNonNull(Object... things) {
-    for (Object thing : things) {
-      checkNonNull(thing);
-    }
-  }
-
-  private void checkNonNullContent(Collection<?> collectionOfThings) {
-    checkNonNull(collectionOfThings);
-    for (Object thing : collectionOfThings) {
-      checkNonNull(thing);
-    }
-  }
-
-  private void addBulkMethodEntriesCount(BulkOps op, long count) {
-    bulkMethodEntries.get(op).add(count);
-  }
-
-  /**
-   * {@inheritDoc}
-   */
-  @Override
-  public Jsr107Cache<K, V> getJsr107Cache() {
-    return jsr107Cache;
+  public Jsr107Cache<K, V> createJsr107Cache() {
+    return new Jsr107CacheImpl();
   }
 
   /**
@@ -631,70 +392,7 @@ public class Ehcache<K, V> implements InternalCache<K, V> {
     return null;
   }
 
-  private final class Jsr107CacheImpl implements Jsr107Cache<K, V> {
-    @Override
-    public void loadAll(Set<? extends K> keys, boolean replaceExistingValues, Function<Iterable<? extends K>, Map<K, V>> loadFunction) {
-      if(keys.isEmpty()) {
-        return ;
-      }
-      if (replaceExistingValues) {
-        loadAllReplace(keys, loadFunction);
-      } else {
-        loadAllAbsent(keys, loadFunction);
-      }
-    }
-
-    @Override
-    public Iterator<Entry<K, V>> specIterator() {
-      return new SpecIterator<>(this, store);
-    }
-
-    @Override
-    public V getNoLoader(K key) {
-      return Ehcache.this.getNoLoader(key);
-    }
-
-    @Override
-    public Map<K, V> getAll(Set<? extends K> keys) {
-      return Ehcache.this.getAllInternal(keys, false);
-    }
-
-    private void loadAllAbsent(Set<? extends K> keys, final Function<Iterable<? extends K>, Map<K, V>> loadFunction) {
-      try {
-        store.bulkComputeIfAbsent(keys, absentKeys -> cacheLoaderWriterLoadAllForKeys(absentKeys, loadFunction).entrySet());
-      } catch (StoreAccessException e) {
-        throw newCacheLoadingException(e);
-      }
-    }
-
-    Map<K, V> cacheLoaderWriterLoadAllForKeys(Iterable<? extends K> keys, Function<Iterable<? extends K>, Map<K, V>> loadFunction) {
-      try {
-        Map<? super K, ? extends V> loaded = loadFunction.apply(keys);
-
-        // put into a new map since we can't assume the 107 cache loader returns things ordered, or necessarily with all the desired keys
-        Map<K, V> rv = new LinkedHashMap<>();
-        for (K key : keys) {
-          rv.put(key, loaded.get(key));
-        }
-        return rv;
-      } catch (Exception e) {
-        throw newCacheLoadingException(e);
-      }
-    }
-
-    private void loadAllReplace(Set<? extends K> keys, final Function<Iterable<? extends K>, Map<K, V>> loadFunction) {
-      try {
-        store.bulkCompute(keys, entries -> {
-          Collection<K> keys1 = new ArrayList<>();
-          for (Map.Entry<? extends K, ? extends V> entry : entries) {
-            keys1.add(entry.getKey());
-          }
-          return cacheLoaderWriterLoadAllForKeys(keys1, loadFunction).entrySet();
-        });
-      } catch (StoreAccessException e) {
-        throw newCacheLoadingException(e);
-      }
-    }
+  private final class Jsr107CacheImpl extends Jsr107CacheBase {
 
     @Override
     public void compute(K key, final BiFunction<? super K, ? super V, ? extends V> computeFunction,
@@ -799,158 +497,6 @@ public class Ehcache<K, V> implements InternalCache<K, V> {
       putObserver.end(PutOutcome.PUT);
       return returnValue;
     }
-
-    @Override
-    public boolean remove(K key) {
-      return removeInternal(key);
-    }
-
-    @Override
-    public void removeAll() {
-      Store.Iterator<Entry<K, ValueHolder<V>>> iterator = store.iterator();
-      while (iterator.hasNext()) {
-        try {
-          Entry<K, ValueHolder<V>> next = iterator.next();
-          remove(next.getKey());
-        } catch (StoreAccessException cae) {
-          // skip
-        }
-      }
-    }
-  }
-
-  private class CacheEntryIterator implements Iterator<Entry<K, V>> {
-
-    private final Store.Iterator<Entry<K, Store.ValueHolder<V>>> iterator;
-    private final boolean quiet;
-    private Cache.Entry<K, ValueHolder<V>> current;
-    private Cache.Entry<K, ValueHolder<V>> next;
-    private StoreAccessException nextException;
-
-    public CacheEntryIterator(boolean quiet) {
-      this.quiet = quiet;
-      this.iterator = store.iterator();
-      advance();
-    }
-
-    private void advance() {
-      try {
-        while (iterator.hasNext()) {
-          next = iterator.next();
-          if (getNoLoader(next.getKey()) != null) {
-            return;
-          }
-        }
-        next = null;
-      } catch (RuntimeException re) {
-        nextException = new StoreAccessException(re);
-        next = null;
-      } catch (StoreAccessException cae) {
-        nextException = cae;
-        next = null;
-      }
-    }
-
-    @Override
-    public boolean hasNext() {
-      statusTransitioner.checkAvailable();
-      return nextException != null || next != null;
-    }
-
-    @Override
-    public Entry<K, V> next() {
-      if (!hasNext()) {
-        throw new NoSuchElementException();
-      }
-
-      if (!quiet) getObserver.begin();
-      if (nextException == null) {
-        if (!quiet) getObserver.end(org.ehcache.core.statistics.CacheOperationOutcomes.GetOutcome.HIT);
-        current = next;
-        advance();
-        return new ValueHolderBasedEntry<>(current);
-      } else {
-        if (!quiet) getObserver.end(org.ehcache.core.statistics.CacheOperationOutcomes.GetOutcome.FAILURE);
-        StoreAccessException cae = nextException;
-        nextException = null;
-        return resilienceStrategy.iteratorFailure(cae);
-      }
-    }
-
-    @Override
-    public void remove() {
-      statusTransitioner.checkAvailable();
-      if (current == null) {
-        throw new IllegalStateException("No current element");
-      }
-      Ehcache.this.remove(current.getKey(), current.getValue().value());
-      current = null;
-    }
-  }
-
-  private static <K> RecoveryCache<K> recoveryCache(final Store<K, ?> store) {
-    return new RecoveryCache<K>() {
-
-      @Override
-      public void obliterate() throws StoreAccessException {
-        store.clear();
-      }
-
-      @Override
-      public void obliterate(K key) throws StoreAccessException {
-        store.remove(key);
-      }
-
-      @Override
-      public void obliterate(Iterable<? extends K> keys) throws StoreAccessException {
-        for (K key : keys) {
-          obliterate(key);
-        }
-      }
-    };
-  }
-
-  private static <K, V> boolean newValueAlreadyExpired(Logger logger, ExpiryPolicy<? super K, ? super V> expiry, K key, V oldValue, V newValue) {
-    if (newValue == null) {
-      return false;
-    }
-
-    Duration duration;
-    try {
-      if (oldValue == null) {
-          duration = expiry.getExpiryForCreation(key, newValue);
-      } else {
-          duration = expiry.getExpiryForUpdate(key, supplierOf(oldValue), newValue);
-      }
-    } catch (RuntimeException re) {
-      logger.error("Expiry computation caused an exception - Expiry duration will be 0 ", re);
-      return true;
-    }
-
-    if (duration != null && duration.isNegative()) {
-      return true;
-    } else {
-      return Duration.ZERO.equals(duration);
-    }
-  }
-
-  private static class ValueHolderBasedEntry<K, V> implements Cache.Entry<K, V> {
-    private final Cache.Entry<K, ValueHolder<V>> storeEntry;
-
-    ValueHolderBasedEntry(Cache.Entry<K, Store.ValueHolder<V>> storeEntry) {
-      this.storeEntry = storeEntry;
-    }
-
-    @Override
-    public K getKey() {
-      return storeEntry.getKey();
-    }
-
-    @Override
-    public V getValue() {
-      return storeEntry.getValue().value();
-    }
-
   }
 
   // The compute function that will return the keys to their NEW values, taking the keys to their old values as input;
@@ -999,7 +545,7 @@ public class Ehcache<K, V> implements InternalCache<K, V> {
     }
 
     private boolean newValueAlreadyExpired(K key, V oldValue, V newValue) {
-      return Ehcache.newValueAlreadyExpired(logger, expiry, key, oldValue, newValue);
+      return EhcacheBase.newValueAlreadyExpired(logger, expiry, key, oldValue, newValue);
     }
 
     public AtomicInteger getActualPutCount() {
